Resetting the card includes sharedsecret
The sharedsecret for communication is a secret to be stored either in mind
or in a secured server(e.g. cloud server). This secret is used for setting
up encrypted data link between a terminal and a card. It's also the salt
in encrypting the actual data(for which the user gives another password).